+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.4.1] - 2026-04-19
+
+### Fixes
+
+* smart copy improvements
+    - CiderPress suffix works for DOS 3.x
+    - fix some issues with filename mapping
+    - skip Apple language detection on non-Apple disks
+* better detection of random access text files
+
 ## [4.4.0] - 2026-04-11
 
 ### Fixes